Code - Web Application/cloud_functions/dataset_consolidator/requirements.txt
pandas
numpy
google-cloud-storage
requests
google-cloud-pubsub
google-cloud-aiplatform
google-cloud-secret-manager